+++ PgBuilder/AbilityModifierPercent.cs
@@ -4,7 +4,7 @@
 
     public class AbilityModifierPercent : AbilityModifier
     {
-        public AbilityModifierPercent(string name, int defaultValue, AbilityBaseValueGetter getter, Func<double, string> displayHandler)
+        public AbilityModifierPercent(string name, int defaultValue, AbilityBaseValueGetter getter, Func<float, string> displayHandler)
             : base(name, getter, displayHandler)
         {
             DefaultValue = defaultValue;
